added user context so user can be global
